fix(kraken): support fractional crypto quantities and improve trade capture
- Change TradeMarker.quantity from int to float for sub-1.0 crypto amounts
- Use float() instead of int(float()) when building trade markers
- Switch Kraken OHLC tick_type from "spot" to "trade" for real volume data
- Append .KRAKEN venue to bare symbols in backtest detail configuration
- Add exc_info=True to trade capture error handlers for better diagnostics

+++ src/api/models/chart_trades.py
@@ -42,7 +42,7 @@ class TradeMarker(BaseModel):
     time: str = Field(..., description="ISO 8601 date format")
     side: str = Field(..., description="Trade side (buy/sell)")
     price: float = Field(..., gt=0, description="Execution price")
-    quantity: int = Field(..., gt=0, description="Trade quantity")
+    quantity: float = Field(..., gt=0, description="Trade quantity")
     pnl: float = Field(..., description="Realized P&L")
